requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-autoslug==1.9.9
django-cors-headers==4.2.0
django-embed-video==1.4.9
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Markdown==3.4.4
mysqlclient==2.2.0
Pillow==10.0.0
psycopg2==2.9.6
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tornado==6.3.2
tzdata==2023.3
urllib3==2.0.3
watchdog==3.0.0
websocket-client==1.6.1
